chore(admin-ui): move <PreviewDataContent /> component under the components folder
